Replaced "ensure_loaded?" with "ensure_compiled" to solve race condition.
Apparently (as by the docs), when executing in interpreted mode
compilition of modules happens in parallel to loading them. So there is
a chance that a module isn't compiled yet and therefore cant be loaded
either. This mirrors an error where _sometimes_ the module can't be
laoded because the file isn't found. Using ensure_compile is said to
solve this exact problem and produces better results.